Split test cases into individual files
