+++ manifest.toml
@@ -16,3 +16,10 @@ instances = { min = 3, max = 100, default = 3 }
 
   [testcases.params]
   latency = { type = "int", desc = "latency between peers", unit = "ms", default = 100 }
+
+[[testcases]]
+name = "eclipse-attack-monopolizing-connections"
+instances = { min = 3, max = 100, default = 3 }
+
+  [testcases.params]
+  latency = { type = "int", desc = "latency between peers", unit = "ms", default = 100 }
